update heatmap on change of species
